# Faster lexical analyzer