Extracted from #2652 

This PR improves the data structure of lexer in RBS.

It improves the parsing performance from ~`14 i/s` to ~`16 i/s` measured by `bin/benchmark-parse.rb`.

<details>
I have `gem_rbs_collection` repository too to load `activerecord` rbs files.

### Baseline

```
➜  rbs git:(38724282) bundle exec ruby bin/benchmark-parse.rb core/**/*.rbs ../../ruby/gem_rbs_collection/gems/activerecord/8.0/*.rbs sig/**/*.rbs
Benchmarking parsing 177 files...
✅ 14.506 i/s (±0.000%)
```

### Fix lexer

```
➜  rbs git:(fix-lexer) bundle exec ruby bin/benchmark-parse.rb core/**/*.rbs ../../ruby/gem_rbs_collection/gems/activerecord/8.0/*.rbs sig/**/*.rbs
Benchmarking parsing 177 files...
✅ 16.667 i/s (±0.000%)
```
</details>